Restructured the component files
