+++ components/OpenThermostatDht.cpp
